+++ CHANGES.md
@@ -1,8 +1,7 @@
 #### unreleased changes
 
-* New features:
-* Bug fixes:
-* Under the hood:
+* Added support for passkeys that was just added in WordFence 9.0.
+* Added a Security tab to the UM Accounts page which contains the 2FA and Passkey setup if it's enabled, so end users can actually set them up.
 
 #### 1.0.2 / 2026-07-13